# swagger-conta-v1.yaml
openapi: 3.0.3
info:
  title: vasco-bank-conta - OpenAPI 3.0
  description: ''
  version: 1.0.11

servers:
  - url: https:/vasco-bank-conta/api/v1
tags:
  - name:  contaCorrente
    description: acesso a conta corrente

paths:

  /contaCorrente/{idUser}:
    post:
      tags:
        - contaCorrente
      summary: Criação da conta
      description: Returns um map da conta corrente do usuario
      operationId: postContaCorrente
      parameters:
        - name: idUser
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contaCorrenteCriar'
          application/xml:
            schema:
              $ref: '#/components/schemas/contaCorrenteCriar'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/contaCorrenteCriar'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contaCorrenteCriar'          
            application/xml:
              schema:
                $ref: '#/components/schemas/contaCorrenteCriar'
        '400':
          description: Invalid ID supplied
        '404':
          description: conta not found
      security:
        - api_key: []
        
    put:
      tags:
        - contaCorrente
      summary: Criação da conta
      description: Returns um map da conta corrente do usuario
      operationId: putContaCorrente
      parameters:
        - name: idUser
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/contaCorrenteCriar'
          application/xml:
            schema:
              $ref: '#/components/schemas/contaCorrenteCriar'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/contaCorrenteCriar'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contaCorrenteCriar'          
            application/xml:
              schema:
                $ref: '#/components/schemas/contaCorrenteCriar'
        '400':
          description: Invalid ID supplied
        '404':
          description: conta not found
      security:
        - api_key: []

    get:
      tags:
        - contaCorrente
      summary: Retorna a conta corrente do usuario
      description: Returns um map da conta corrente do usuario
      operationId: getContaCorrentById
      parameters:
        - name: idUser
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contaCorrente'          
            application/xml:
              schema:
                $ref: '#/components/schemas/contaCorrente'
        '400':
          description: Invalid ID supplied
        '404':
          description: conta not found
      security:
        - api_key: []
        
    delete:
      tags:
        - contaCorrente
      summary: Retorna a conta corrente do usuario
      description: Returns um map da conta corrente do usuario
      operationId: deleteContaCorrentById
      parameters:
        - name: idUser
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contaCorrente'          
            application/xml:
              schema:
                $ref: '#/components/schemas/contaCorrente'
        '400':
          description: Invalid ID supplied
        '404':
          description: conta not found
      security:
        - api_key: []
        
  /contaCorrente/saldo/{idConta}:
    get:
      tags:
        - contaCorrente
      summary: Retorna o saldo da conta
      description: Place a new order in the store
      operationId: getSaldo
      parameters:
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
   
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saldo'
        '405':
          description: Invalid input
          
    put:
      tags:
        - contaCorrente
      summary: Atualiza o saldo da conta
      description: Place a new order in the store
      operationId: putSaldosaldo
      parameters:
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saldo'
          application/xml:
            schema:
              $ref: '#/components/schemas/saldo'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/saldo'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saldo'
        '405':
          description: Invalid input
          
  /contaCorrente/cheque/{idConta}:
    post:
      tags:
        - contaCorrente
      summary: Cria um cheque na conta
      description: Returns um map da conta corrente do usuario
      operationId: postCheque
      parameters:
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/cheque'
          application/xml:
            schema:
              $ref: '#/components/schemas/cheque'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/cheque'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/contaCorrente'          
            application/xml:
              schema:
                $ref: '#/components/schemas/contaCorrente'
        '400':
          description: Invalid ID supplied
        '404':
          description: conta not found
      security:
        - api_key: []
        
  /contaCorrente/cheque/{idConta}/{idCheque}:
    get:
      tags:
        - contaCorrente
      summary: Retorna o cheque da conta
      description: Returns um map da conta corrente do usuario
      operationId: getCheque
      parameters:
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: idCheque
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cheque'          
            application/xml:
              schema:
                $ref: '#/components/schemas/cheque'
        '400':
          description: Invalid ID supplied
        '404':
          description: conta not found
      security:
        - api_key: []
        
  /contaCorrente/cheque/all/{idConta}:
    get:
      tags:
        - contaCorrente
      summary: Retorna todos os cheques da conta
      description: Returns um map da conta corrente do usuario
      operationId: getCheques
      parameters:
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64

      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cheques'          
            application/xml:
              schema:
                $ref: '#/components/schemas/cheques'
        '400':
          description: Invalid ID supplied
        '404':
          description: conta not found
      security:
        - api_key: []
          
  /contaCorrente/cheque/{idConta}/{idCheque}/limite:
    get:
      tags:
        - contaCorrente
      summary: retorna o limite do cheque da conta
      description: Place a new order in the store
      operationId: getChequeLimite
      parameters:
        - name: idCheque
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saldo'
        '405':
          description: Invalid input
    put:
      tags:
        - contaCorrente
      summary: Atualiza o limite do cheque da conta
      description: Place a new order in the store
      operationId: chequepust
      parameters:
        - name: idCheque
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/cheque'
          application/xml:
            schema:
              $ref: '#/components/schemas/cheque'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/cheque'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/cheque'
        '405':
          description: Invalid input
          
  /contaCorrente/cheque/{idConta}/{idCheque}/juros:
    get:
      tags:
        - contaCorrente
      summary: Retorna os juros do cheque da conta
      description: Place a new order in the store
      operationId: chequejuros
      parameters:
        - name: idCheque
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saldo'
        '405':
          description: Invalid input
          

    put:
      tags:
        - contaCorrente
      summary: Atualiza os juros do cheque da conta
      description: Place a new order in the store
      operationId: chequeput
      parameters:
        - name: idCheque
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
        - name: idConta
          in: path
          description: ID do usuario
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/saldo'
          application/xml:
            schema:
              $ref: '#/components/schemas/saldo'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/saldo'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/saldo'
        '405':
          description: Invalid input
          
  /contaCorrente/extrato/{idConta}:
    get:
      tags:
        - contaCorrente
      summary: retorna o extrato da conta
      description: For valid response try integer IDs with value <= 5 or > 10. Other values will generate exceptions.
      operationId: getOrderById
      parameters:
        - name: idConta
          in: path
          description: ID of order that needs to be fetched
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/extrato'          
            application/xml:
              schema:
                $ref: '#/components/schemas/extrato'
        '400':
          description: Invalid ID supplied
        '404':
          description: Order not found

 
    
components:
  schemas:
    contaCorrente:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        nome:
          type: string
          example: 'John pascal'
        primeiroNome:
          type: string
          example: 'John'
        agencia:
          type: string
          example: '1232'
        numero:
          type: string
          example: '23232'
        digito:
          type: string
          example: '3'
        numeroComDigito:
          type: string
          example: '23232-3'
        dataNascimento:
          type: string
          example: "2023-06-06"
        cpf:
          type: string
          example: '999.999.999-99'
        profissao:
          type: string
          example: 'Desenvolvedor'
        elegivelCredito:
          type: boolean
          example: 'false'
        saldo:
          type: number
          example: 0
        cheque:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/cheque'
      xml:
        name: contaCorrente


    contaCorrenteCriar:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        nome:
          type: string
          example: 'John pascal'
        primeiroNome:
          type: string
          example: 'John'
        agencia:
          type: string
          example: '1232'
        numero:
          type: string
          example: '23232'
        digito:
          type: string
          example: '3'
        numeroComDigito:
          type: string
          example: '23232-3'
        dataNascimento:
          type: string
          example: "2023-06-06"
        cpf:
          type: string
          example: '999.999.999-99'
        profissao:
          type: string
          example: 'Desenvolvedor'
        elegivelCredito:
          type: boolean
          example: 'false'
        saldo:
          type: number
          example: 0
        cheque:
          type: array
          nullable: true
          items: 
            $ref: '#/components/schemas/cheque'
      xml:
        name: contaCorrenteCriar        
    cheque:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        valor:
          type: number
          example: 1000
        limite:
          type: number
          example: 1000
        juros:
          type: number
          example: 1000
        valorExtenso:
          type: string
          example: 'mil reais'
        dataAtual:
          type: string
          example: "2023-06-06"
        nome:
          type: string
          example: 'john'
        numero:
          type: string
          example: '123123'
        serie:
          type: string
          example: '123123'
        dataEntrada:
          type: string
          example: "2023-06-06"
  
      xml:
        name: cheque
        
    cheques:
      type: object
      properties:
        cheques:
          type: array
          xml:
            wrapped: true
          items:
            $ref: '#/components/schemas/cheque'
      xml:
        name: cheques
        
    saldo:
      type: object
      properties:
        valor:
          type: number
          format: int64
          example: 99.999
      xml:
        name: saldo
        
    extrato:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        dataOperacao:
          type: string
          example: "2023-06-06"
        operacao:
          type: string
          enum:
            - PIX
            - BOLETO
            - INTRABANCO
            - PAGCARTAO
      xml:
        name: extrato
        
   
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://vasco-bank-conta/api/v1/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header